refactor: consumer parameters select input
