+++ src/components/@shared/FormInput/InputElement/ConsumerParameters/SelectInput.module.css
@@ -1,0 +1,16 @@
+.optionsContainer {
+  position: relative;
+  margin-bottom: calc(var(--spacer) / 2);
+}
+
+.optionsContainer.hasError * {
+  border-color: var(--brand-alert-red);
+}
+
+.hasError label {
+  color: var(--brand-alert-red);
+}
+
+.error {
+  composes: error from '../../index.module.css';
+}
