+++ src/components/@shared/FormInput/InputElement/ConsumerParameters/index.module.css
@@ -1,20 +1,3 @@
 .container {
   border: 1px solid var(--border-color);
 }
-
-.optionsContainer {
-  position: relative;
-  margin-bottom: calc(var(--spacer) / 2);
-}
-
-.optionsContainer.hasError * {
-  border-color: var(--brand-alert-red);
-}
-
-.hasError label {
-  color: var(--brand-alert-red);
-}
-
-.error {
-  composes: error from '../../index.module.css';
-}
